feature: toolbox - add k8s context switching
