+++ infra-scripts/clitools/run_tests.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
 # Run the tests
-python3 -m unittest test_toolbox.py -v 
+# Get directory of this script
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+cd "$SCRIPT_DIR"
+python3 -m unittest test_toolbox.py -v
