+++ infra-scripts/clitools/toolbox/__init__.py
@@ -1,0 +1,3 @@
+"""
+Toolbox package for connecting to PostHog toolbox pod.
+"""
